Cleanup - server config moved into server
